+++ src/Aspire.Dashboard/Aspire.Dashboard.csproj
@@ -174,6 +174,7 @@
       <Link>Protos\resource_service.proto</Link>
     </Protobuf>
     <Compile Include="..\Aspire.Hosting\Extensions\ChannelExtensions.cs" Link="Extensions\ChannelExtensions.cs" />
+    <Compile Include="..\Aspire.Hosting\Utils\EnvironmentUtil.cs" Link="Utils\EnvironmentUtil.cs" />
   </ItemGroup>
 
 </Project>
